feat: client support for image-based patterns

+++ hopfield-neural-network/src/main/java/ua/kpi/iasa/sd/hopfieldneuralnetwork/domain/Weight.java
@@ -1,4 +1,4 @@
 package ua.kpi.iasa.sd.hopfieldneuralnetwork.domain;
 
-public record Weight(float[][] w) {
+public record Weight(int[][] w) {
 }
